[feat] : 나머지 error handling도 구현함

+++ server/src/article/service/article.service.ts
@@ -13,8 +13,10 @@ export class ArticleService {
     return await this.articleRepository.find();
   }
 
-  async getArticleById(id: number): Promise<Article> {
-    return await this.articleRepository.findArticleById(id);
+  async getArticleById(articleId: number): Promise<Article> {
+    const article = await this.articleRepository.findArticleById(articleId);
+    if (!article) throw new HttpException('존재하지 않는 article입니다', 404);
+    return article;
   }
 
   async createArticle(userId: number, dto: CreateArticleDto): Promise<void> {
@@ -28,17 +30,17 @@ export class ArticleService {
     dto: UpdateArticleDto,
   ): Promise<void> {
     const article = await this.articleRepository.findArticleById(articleId);
-    if (article.fk_user_id !== userId) {
-      throw new HttpException('This is not your article.', 401);
-    }
+    if (!article) throw new HttpException('존재하지 않는 article입니다', 404);
+    if (article.fk_user_id !== userId)
+      throw new HttpException('당신의 article이 아닙니다.', 401);
     await this.articleRepository.updateArticle(articleId, dto);
   }
 
   async deleteArticle(userId: number, articleId: number): Promise<void> {
     const article = await this.articleRepository.findArticleById(articleId);
-    if (article.fk_user_id !== userId) {
-      throw new HttpException('This is not your article.', 401);
-    }
+    if (!article) throw new HttpException('존재하지 않는 article입니다', 404);
+    if (article.fk_user_id !== userId)
+      throw new HttpException('당신의 article이 아닙니다.', 401);
     await this.articleRepository.deleteArticle(articleId);
   }
 }
